Before tarring site, set file read permissions
Pages DFS needs the "group read" permission set on any file it serves,
so add read permission before tarring to prevent unreadable files from
being deployed.

This only applies to files that are built on an actions runner, not
files that are cloned from a git repoitory, because git's permissions
are limited to the equivalents of chmod 644 and 755.